TST: Updated `test_fracture_network_*.py`
